Edited JZombiesBuilder to test seed / random generator
- Set Seed to static value (1)
- Added JZombiesBuilder to context in order to enable scheduling
- Created ScheduledMethod step() which generates some random Numbers and
stops the run after 10 ticks. It then prints the addition of all x and y
coordinates as well as the multiplication of some random numbers. This
method is quite random but should enable tracking of the changes in the
position of the Zombies as well as as the current random numbers